Moved DiagrammedExp and DigrammedExprMacro from scalactic package to scalatest package.
