+++ src/main/scala/org/scalatest/DiagrammedAssertionsMacro.scala
@@ -15,7 +15,7 @@
  */
 package org.scalatest
 
-import org.scalactic.{BooleanMacro, DiagrammedExprMacro}
+import org.scalactic.BooleanMacro
 import reflect.macros.Context
 
 private[scalatest] object DiagrammedAssertionsMacro {
